Added animation transitions on navigating from empty state to data state

+++ swift-todo-list/Views/ ListView.swift
@@ -12,19 +12,28 @@ struct ListView: View {
     @EnvironmentObject var viewModel : ListViewModel
     
     var body: some View {
-        List{
-            ForEach(viewModel.items) { item in
-                ListTile(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            viewModel.updateItem(item: item)
-                        }
+        ZStack{
+            
+            if viewModel.items.isEmpty{
+                NoItemsView()
+                    .transition(.opacity.animation(.easeIn))
+            } else {
+                List{
+                    ForEach(viewModel.items) { item in
+                        ListTile(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    viewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete (perform: viewModel.deleteItem)
+                    .onMove (perform: viewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete (perform: viewModel.deleteItem)
-            .onMove (perform: viewModel.moveItem)
+            
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List 📝")
         .toolbar{
             ToolbarItem(placement: .topBarLeading){
